ai-service/requirements.txt
fastapi
uvicorn
easyocr
python-multipart
langchain
google-generativeai